requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fastapi
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hub-sdk==0.0.18
huggingface-hub==0.27.1
huggingface_hub
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
koreanize-matplotlib==0.1.1
lapx>=0.5.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.0.2
omegaconf==2.3.0
openai==1.59.8
opencv-python==4.10.0.84
opencv-python-headless
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
py-cpuinfo==9.0.0
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.1
python-multipart
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette
sympy==1.13.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==0.8.13.dev0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers
tzdata==2024.2
ultralytics==8.1.0
urllib3==2.3.0
uvicorn[standard]
wheel==0.44.0
yarl==1.18.3
yt-dlp==2025.1.15